crop-backend-main/crop-backend-main/requirements.txt
flask>=2.3.2
flask-cors>=4.0.0
werkzeug>=2.3.6
gunicorn>=21.2.0
tensorflow>=2.13.0
numpy>=1.24.3
opencv-python-headless>=4.8.0.74
pillow>=10.0.0
python-multipart>=0.0.6
edge-tts>=6.1.9